cli/requirements.txt
SQLAlchemy==2.0.16
sentence-transformers==2.2.2
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
pinecone-client
sqlite3
Scrapy==2.9.0
openai==0.27.8
click==8.1.3
datasets==2.13.1
